Added radio buttons to page & edited text content.
I set up radio buttons on the page which I will use
in java to allow users to make their grid more
customizable. Each button corresponds to a color
and that will be the color for the hover effect &
if a user clicks to set a color for a square.